fixing logical error

+++ 0x15-file_io/1-create_file.c
@@ -16,7 +16,7 @@ int create_file(const char *filename, char *text_content)
 		return (-1);
 	if (text_content != NULL)
 	{
-		len = sizeof(text_content) / sizeof(char) - 1;
+		len = (int)sizeof(text_content) / sizeof(char) - 1;
 	}
 
 	fo = open(filename, O_CREAT | O_RDWR | O_TRUNC, 0600);
